requirements.txt
aiogram==2.25.1
Flask==2.3.3
gspread==5.12.4
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.104.0
aiohttp==3.8.5
apscheduler==3.10.4
requests==2.31.0
python-dotenv==1.0.0
